+++ Evergreen/Public/Get-MicrosoftPowerToys.ps1
@@ -1,4 +1,4 @@
-Function Get-MicrosoftPowerToys  {
+Function Get-MicrosoftPowerToys {
     <#
         .SYNOPSIS
             Returns the latest Microsoft PowerToys version number and download.
@@ -28,19 +28,17 @@ Function Get-MicrosoftPowerToys  {
     $res = Get-FunctionResource -AppName ("$($MyInvocation.MyCommand)".Split("-"))[1]
     Write-Verbose -Message $res.Name
 
-    # Get latest version and download latest release via GitHub API
-    $iwcParams = @{
-        Uri         = $res.Get.Uri
-        ContentType = $res.Get.ContentType
-    }
-    $Content = Invoke-WebContent @iwcParams
-
-    # Convert the returned release data into a useable object with Version, URI etc.
+    # Pass the repo releases API URL and return a formatted object
     $params = @{
-        Content      = $Content
+        Uri          = $res.Get.Uri
         MatchVersion = $res.Get.MatchVersion
-        VersionTag   = $res.Get.VersionTag
+        Filter       = $res.Get.MatchFileTypes
     }
-    $object = ConvertFrom-GitHubReleasesJson @params
-    Write-Output -InputObject $object
+    $object = Get-GitHubRepoRelease @params
+    If ($object) {
+        Write-Output -InputObject $object
+    }
+    Else {
+        Write-Warning -Message "$($MyInvocation.MyCommand): Failed to return a usable object from the repo."
+    }
 }